+++ src/main/java/security/fixes/JinjaAutoinspectFixer.kt
@@ -39,7 +39,7 @@ class JinjaAutoinspectUnconditionalFixer : LocalQuickFix, IntentionAction, HighP
     fun runFix(project: Project, file: PsiFile, originalElement: PsiElement): PyCallExpression? {
         if (originalElement !is PyCallExpression) return null
         if (file !is PyFile) return null
-        var newEl = originalElement.copy() as PyCallExpression
+        val newEl = originalElement.copy() as PyCallExpression
         val autoescapeArgument = newEl.getKeywordArgument("autoescape")
         val elementGenerator = PyElementGenerator.getInstance(project)
         val newArg = elementGenerator.createKeywordArgument(file.languageLevel, "autoescape", "True")