+++ src/main/java/security/validators/AssertInspection.kt
@@ -10,7 +10,7 @@ import com.jetbrains.python.psi.PyAssertStatement
 import security.Checks
 
 class AssertInspection : PyInspection() {
-    val check = Checks.AssertCheck;
+    val check = Checks.AssertCheck
 
     override fun getStaticDescription(): String? {
         return check.getStaticDescription()
@@ -21,8 +21,7 @@ class AssertInspection : PyInspection() {
                               session: LocalInspectionToolSession): PsiElementVisitor = Visitor(holder, session)
 
     private class Visitor(holder: ProblemsHolder, session: LocalInspectionToolSession) : PyInspectionVisitor(holder, session) {
-        override fun visitPyAssertStatement(node: PyAssertStatement?) {
-            if (node == null) return
+        override fun visitPyAssertStatement(node: PyAssertStatement) {
             if (node.containingFile.name.contains("test"))
                 return
             holder?.registerProblem(node, Checks.AssertCheck.getDescription(), ProblemHighlightType.WEAK_WARNING)