+++ src/main/java/security/validators/DjangoDebugModeSettingsInspection.kt
@@ -9,7 +9,7 @@ import com.jetbrains.python.psi.PyAssignmentStatement
 import security.Checks
 
 class DjangoDebugModeSettingsInspection : PyInspection() {
-    val check = Checks.DjangoDebugModeCheck;
+    val check = Checks.DjangoDebugModeCheck
 
     override fun getStaticDescription(): String? {
         return check.getStaticDescription()
@@ -20,12 +20,12 @@ class DjangoDebugModeSettingsInspection : PyInspection() {
                               session: LocalInspectionToolSession): PsiElementVisitor = Visitor(holder, session)
 
     private class Visitor(holder: ProblemsHolder, session: LocalInspectionToolSession) : PyInspectionVisitor(holder, session) {
-        override fun visitPyAssignmentStatement(node: PyAssignmentStatement?) {
-            if (node?.containingFile?.name != "settings.py") return;
+        override fun visitPyAssignmentStatement(node: PyAssignmentStatement) {
+            if (node.containingFile?.name != "settings.py") return
             val leftExpression = node.leftHandSideExpression?.text ?: return
-            if (leftExpression != "DEBUG") return;
+            if (leftExpression != "DEBUG") return
             val assignedValue = node.assignedValue ?: return
-            if (assignedValue.textMatches("True").not()) return;
+            if (assignedValue.textMatches("True").not()) return
             holder?.registerProblem(node, Checks.DjangoDebugModeCheck.getDescription())
         }
     }