+++ src/main/java/security/validators/DjangoSafeStringInspection.kt
@@ -1,17 +1,17 @@
 package security.validators
 
 import com.intellij.codeInspection.LocalInspectionToolSession
+import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import com.jetbrains.python.inspections.PyInspection
 import com.jetbrains.python.inspections.PyInspectionVisitor
 import com.jetbrains.python.psi.PyCallExpression
 import security.Checks
-import security.fixes.TempfileMksFixer
-import security.helpers.QualifiedNames.getQualifiedName
+import security.helpers.QualifiedNames
 
-class TempfileMktempInspection : PyInspection() {
-    val check = Checks.TempfileMktempCheck
+class DjangoSafeStringInspection : PyInspection() {
+    val check = Checks.DjangoSafeStringCheck
 
     override fun getStaticDescription(): String? {
         return check.getStaticDescription()
@@ -22,12 +22,13 @@ class TempfileMktempInspection : PyInspection() {
                               session: LocalInspectionToolSession): PsiElementVisitor = Visitor(holder, session)
 
     private class Visitor(holder: ProblemsHolder, session: LocalInspectionToolSession) : PyInspectionVisitor(holder, session) {
+        val methodNames = arrayOf("SafeString", "mark_safe", "SafeBytes", "SafeUnicode", "SafeText")
         override fun visitPyCallExpression(node: PyCallExpression) {
             val calleeName = node.callee?.name ?: return
-            if (calleeName != "mktemp") return
-            val qualifiedName = getQualifiedName(node) ?: return
-            if (qualifiedName != "tempfile.mktemp") return
-            holder?.registerProblem(node, Checks.TempfileMktempCheck.getDescription(), TempfileMksFixer())
+            if (!listOf(*methodNames).contains(calleeName)) return
+            val qualifiedName = QualifiedNames.getQualifiedName(node) ?: return
+            if (!qualifiedName.startsWith("django.utils.safestring")) return
+            holder?.registerProblem(node, Checks.DjangoSafeStringCheck.getDescription(), ProblemHighlightType.WEAK_WARNING)
         }
     }
 }