+++ src/main/java/security/validators/SqlInjectionInspection.kt
@@ -9,7 +9,7 @@ import com.jetbrains.python.psi.*
 import security.Checks
 
 class SqlInjectionInspection : PyInspection() {
-    val check = Checks.SqlInjectionCheck;
+    val check = Checks.SqlInjectionCheck
 
     override fun getStaticDescription(): String? {
         return check.getStaticDescription()
@@ -23,7 +23,7 @@ class SqlInjectionInspection : PyInspection() {
         // Double-word SQL commands (high-certainty)
         val certainlySqlStartingStrings = arrayOf("INSERT INTO ", "DELETE FROM", "ALTER TABLE ", "DROP DATABASE ", "CREATE DATABASE ")
         // Double-word SQL commands (low-certainty)
-        val possiblySqlCommandPairs = mapOf<String, String>("SELECT " to " FROM ", "UPDATE " to " SET ")
+        val possiblySqlCommandPairs = mapOf("SELECT " to " FROM ", "UPDATE " to " SET ")
 
         fun looksLikeSql(str: String) : Boolean {
             // Quickly respond to double-worded SQL statements
@@ -37,15 +37,13 @@ class SqlInjectionInspection : PyInspection() {
             return false
         }
 
-        override fun visitPyFormattedStringElement(node: PyFormattedStringElement?) {
+        override fun visitPyFormattedStringElement(node: PyFormattedStringElement) {
             // F-string
-            if (node == null) return
             if (!looksLikeSql(node.content)) return
             holder?.registerProblem(node, Checks.SqlInjectionCheck.getDescription())
         }
 
-        override fun visitPyStringLiteralExpression(node: PyStringLiteralExpression?) {
-            if (node == null) return
+        override fun visitPyStringLiteralExpression(node: PyStringLiteralExpression) {
             if (!looksLikeSql(node.stringValue)) return
 
             // .Format() string